Refactor a bit to avoid redisio dependency on in front-end
